==> Codes/3_LCA analysis based on the new LCI database/build_other_CFmatrix/02_build_countryCF_IPCC_Recipe_PSL_AWARE_mat_for_regLCA.ipynb <==
# Build characterization matrices <br>
- prepared regionalized CF
- non-regionalized CF from ecoinvent LCIA

import pickle
import collections
import numpy as np
import scipy.sparse as sp
import pandas as pd
import brightway2 as bw
import hashlib

!uname -a

!ls ../data/final_matrix/

with open('../data/final_matrix/full_index.p', 'rb') as i:
    full_index = pickle.load(i)
with open('../data/final_matrix/biosphere_index.p', 'rb') as i:
    bio_index = pickle.load(i)

# regionalized CF.
CF_aware = pd.read_csv('parse_CF_data/prepared_characterization_factors/prepared_aware_CF.csv', index_col=0,keep_default_na=False)
CF_lu_op = pd.read_csv('parse_CF_data/prepared_characterization_factors/prepared_land_use_occupation_CF.csv', index_col=0,keep_default_na=False)
CF_lu_tf = pd.read_csv('parse_CF_data/prepared_characterization_factors/prepared_land_use_transformation_CF.csv', index_col=0,keep_default_na=False)

%%time
bw.projects.set_current('regeco')
db = bw.Database('ecoinvent 3.7.1_cutoff_no_marketgroups')

## IPCC

[m for m in bw.methods if m[0].lower().startswith('ipcc 2013')]

ipcc = ('IPCC 2013', 'climate change', 'GWP 100a')

%%time
lca = bw.LCA({db.random():1}, method = ipcc)
lca.lci()
lca.lcia()

Current study: bio_index in biosphere matrix and inventory matrix is sorted by bio codes<br> 
Brightway: biosphere, inventory, and characterization matrix are ordered as in lca.biosphere_dict. <br> therefore when we build CF matrix, we need to reorder characterization matrix.

lca.biosphere_dict

len(bio_index) == len(lca.biosphere_dict)#('biosphere3', '38a622c6-f086-4763-a952-7c6b3b1c42ba'): 0,

biosphere_dict_list = sorted([(k[1], v) for k,v in lca.biosphere_dict.items()], key=lambda x:x[0])

[e[0] for e in biosphere_dict_list] == bio_index

reordered_index = np.array([e[1] for e in biosphere_dict_list])
reordered_index.shape

ipcc_cf_mat = sp.diags(lca.characterization_matrix.diagonal()[reordered_index]).tocsr()
ipcc_cf_mat.shape

## ReCiPe

[m for m in bw.methods if m[0].lower().startswith('recipe endpoint')] #all with '(obsolete)'

recipe = ('ReCiPe Endpoint (H,A)', 'total', 'total')

lca.switch_method(recipe)
lca.redo_lcia()
lca.characterization_matrix

new_biosphere_dict_list = sorted([(k[1], v) for k,v in lca.biosphere_dict.items()], key=lambda x:x[0])
biosphere_dict_list == new_biosphere_dict_list

recipe_cf_mat = sp.diags(lca.characterization_matrix.diagonal()[reordered_index]).tocsr()
recipe_cf_mat

## AWARE

CF_aware.drop("World average",axis=0,inplace=True)

CF_aware.shape

CF_aware.head()

aware_region_cf_dict = {}
not_available = set() #bio_flow in aware but not used by our bio_index
for region, cfs in CF_aware.iterrows():
    cf_array = np.zeros(len(bio_index))
    for i, val in cfs.iteritems():
        if i in bio_index:
            cf_array[bio_index.index(i)] = val
        else:
            not_available.add(i)
    aware_region_cf_dict[region] = cf_array
not_available    

bio = bw.Database('biosphere3').load()
for na in not_available:
    print(bio.get(('biosphere3', na))['name'], '|', bio.get(('biosphere3', na))['categories'])

#Stack arrays in sequence vertically (row wise). T. (len(bio_index), len(full_index) regions)
%time aware_cf_mat = sp.csc_matrix(np.vstack(aware_region_cf_dict[full_i[0]] for full_i in full_index)).T
aware_cf_mat

np.array_equal(np.flatnonzero(np.array(aware_cf_mat.sum(axis=1))), np.flatnonzero(cf_array))

construct a aware_regCF_mat for 144 countries, in the shape of 144*144.

co_l = sorted(list({i[0][:2] if i[0] != 'Canada without Quebec' else 'CA'  for i in full_index}))
CF_aware_co_dict = abs(CF_aware[CF_aware.columns[0]])
CF_aware_reg_144co = [CF_aware_co_dict[c] for c in co_l]
CF_aware_reg_144co_mat = np.diag(CF_aware_reg_144co)
CF_aware_reg_144co_mat

CF_aware_co_dict['CY']

# PSLglobal

print(CF_lu_op.shape)
print(CF_lu_tf.shape)
CF_lu_op.head()

CF_lu = pd.concat([CF_lu_op,CF_lu_tf],axis=1)
CF_lu.to_csv(f'parse_CF_data/prepared_characterization_factors/prepared_land_use_CF.csv')
CF_lu.drop("World average",axis=0,inplace=True)
CF_lu.shape

pslglo_region_cf_dict = {}
not_available = set()
for region, cfs in CF_lu.iterrows():
    cf_array = np.zeros(len(bio_index))
    for i, val in cfs.iteritems():
        if i in bio_index:
            cf_array[bio_index.index(i)] = val
        else:
            not_available.add(i)
    pslglo_region_cf_dict[region] = cf_array
len(not_available)

for na in not_available:
    print(bio.get(('biosphere3', na))['name'])

%time pslglo_cf_mat = sp.csc_matrix(np.vstack(pslglo_region_cf_dict[full_i[0]] for full_i in full_index)).T

pslglo_cf_mat

np.array_equal(np.flatnonzero(np.array(pslglo_cf_mat.sum(axis=1))), np.flatnonzero(cf_array))

# Save

with open('../data/final_matrix/CF_ipcc.p', 'wb') as o:
    pickle.dump(ipcc_cf_mat, o)

with open('../data/final_matrix/CF_recipe.p', 'wb') as o:
    pickle.dump(recipe_cf_mat, o)

with open('../data/final_matrix/CF_aware.p', 'wb') as o:
    pickle.dump(aware_cf_mat, o)

pslglo_cf_mat.data.nbytes*1e-6

with open('../data/final_matrix/CF_pslglo.p', 'wb') as o:
    pickle.dump(pslglo_cf_mat, o)

with open('../data/final_matrix/CF_aware_reg_144co_mat.p', 'wb') as o:
    pickle.dump(CF_aware_reg_144co_mat, o)

!ls -lh ../data/final_matrix
