==> Codes/3_LCA analysis based on the new LCI database/build_other_CFmatrix/03_build_worldCF_PSL_AWARE_mat_for_ecoinventLCA.ipynb <==
import pickle
import collections
import pandas as pd
import numpy as np
import scipy.sparse as sp
import brightway2 as bw

to get lca.biosphere_dict

bw.projects.set_current('regeco')
db = bw.Database('ecoinvent 3.7.1_cutoff_no_marketgroups')
lca = bw.LCA({db.random(): 1}, bw.methods.random())
lca.lci()

psl

pslglo_cfs = pd.read_csv('parse_CF_data/prepared_characterization_factors/prepared_land_use_CF.csv', index_col=0,keep_default_na=False)
pslglo_cf_dict = pslglo_cfs.loc['World average'].to_dict()
pslglo_cf_dict

pslglo_cf_arr_eco = np.array([pslglo_cf_dict.get(key[1],0) for key, pos in lca.biosphere_dict.items()])
pslglo_cf_mat_eco = sp.diags(pslglo_cf_arr_eco).tocsr()
pslglo_cf_mat_eco

with open('../data/final_matrix/CF_pslglo_glo_eco.p', 'wb') as o:
    pickle.dump(pslglo_cf_mat_eco, o)

aware

aware_cfs = pd.read_csv('parse_CF_data/prepared_characterization_factors/prepared_aware_CF.csv', index_col=0,keep_default_na=False)
aware_cf_dict = aware_cfs.loc['World average'].T.to_dict()

aware_cf_arr_eco = np.array([aware_cf_dict.get(key[1],0) for key, pos in lca.biosphere_dict.items()])
aware_cf_mat_eco = sp.diags(aware_cf_arr_eco).tocsr()
aware_cf_mat_eco

with open('../data/final_matrix/CF_aware_glo_eco.p', 'wb') as o:
    pickle.dump(aware_cf_mat_eco, o)

bio_inv_dict = {v:k for k,v in lca.biosphere_dict.items()}

for i, v in enumerate(aware_cf_arr_eco):
    if v:
        b = bw.get_activity(bio_inv_dict[i])
        print(b)
