==> Codes/3_LCA analysis based on the new LCI database/build_other_CFmatrix/04_build_worldCF_PSL_AWARE_mat_for_regLCA.ipynb <==
import pickle
import collections
import numpy as np
import scipy.sparse as sp
import pandas as pd
import brightway2 as bw
import hashlib

### build the CF_glo matrix based on regional CF file

with open('../data/final_matrix/biosphere_index.p', 'rb') as i:
    bio_index = pickle.load(i)

aware_cfs = pd.read_csv('parse_CF_data/prepared_characterization_factors/prepared_aware_CF.csv', index_col=0,keep_default_na=False)
aware_cf_dict = aware_cfs.loc['World average'].T.to_dict()
len(aware_cf_dict)

aware_cf_array = np.array([aware_cf_dict.get(i,0) for i in bio_index])
aware_cf_mat = sp.diags(aware_cf_array).tocsr()
aware_cf_mat

with open('../data/final_matrix/CF_aware_glo.p', 'wb') as o:
    pickle.dump(aware_cf_mat, o)

pslglo_cfs = pd.read_csv('parse_CF_data/prepared_characterization_factors/prepared_land_use_CF.csv', index_col=0,keep_default_na=False)
pslglo_cf_dict = pslglo_cfs.loc['World average'].to_dict()
len(pslglo_cf_dict)

pslglo_cf_array = np.array([pslglo_cf_dict.get(i,0) for i in bio_index])
pslglo_cf_mat = sp.diags(pslglo_cf_array).tocsr()
pslglo_cf_mat

with open('../data/final_matrix/CF_pslglo_glo.p', 'wb') as o:
    pickle.dump(pslglo_cf_mat, o)
